Added date information
I added the functionality to enter a date that we will display in
addition to the temperature conversion. Note the handling of the input
stream in line 11. Since nextInt() leaves a trailing new line character,
we must absorb it with a call to nextLine() to be able to read the date
information from the user.