Fixed the error in Bad2.java
The original Bad2.java file contained an instance in which we divided by
zero, which produced a runtime error. To correct the error, I set the
variable "b" to a non-zero integer. Specifically, I chose the value 11,
since it is one of the divisors of the variable "a", which is equal to
the value 1331.