Fixed the error in Bad3.java
I corrected the error in the Bad3.java file, which was a mismatch
between the data type of the variable "letter" and the value the
original file attempted to assign to it. The data type was String, but
the literal was a character due to its enclosure within single quotes.
To correct the error, I replaced the single quotes with double quotes,
making the literal a String. Alternatively, I could have changed the
data type of the variable "letter" to char.